Changes to URI for participant-requests

+++ server/src/main/java/server/api/ParticipantController.java
@@ -9,7 +9,7 @@ import server.database.ParticipantRepository;
 import java.util.List;
 
 @RestController
-@RequestMapping("/api/participants")
+@RequestMapping("/api/events/{eventId}/participants")
 public class ParticipantController {
     private final ParticipantRepository repo;
 
@@ -24,15 +24,17 @@ public class ParticipantController {
     /**
      * Mapping for adding a participant
      * @param participant the participant to add
+     * @param eventId the event that the participant belongs to
      * @return the participant response entity
      */
     @PostMapping
     @ResponseStatus(HttpStatus.OK)
     public ResponseEntity<Participant>
-        add(@RequestBody Participant participant) {
+        add(@RequestBody Participant participant,
+            @PathVariable("eventId") long eventId) {
         if (participant.getEvent() == null ||
                 isNullOrEmpty(participant.getName()) ||
-                isNullOrEmpty(participant.getEvent().getEventName())) {
+                participant.getEvent().getId() != eventId) {
             return ResponseEntity.badRequest().build();
         }
 
@@ -48,22 +50,31 @@ public class ParticipantController {
      * Mapping for getting a participant by id
      * @param id the id of the participant to get
      * @return the participant found or a bad request if no participant
+     * @param eventId the event that the participant belongs to
      * can be found
      */
     @GetMapping("/{id}")
-    public ResponseEntity<Participant> getById(@PathVariable("id") long id) {
+    public ResponseEntity<Participant>
+        getById(@PathVariable("id") long id,
+                @PathVariable("eventId") long eventId) {
         if (id < 0 || !repo.existsById(id)) {
             return ResponseEntity.badRequest().build();
         }
-        return ResponseEntity.ok(repo.findById(id).get());
+
+        Participant participant = repo.findById(id).get();
+        if (participant.getEvent().getId() != eventId)
+            return ResponseEntity.badRequest().build();
+
+        return ResponseEntity.ok(participant);
     }
 
     /**
      * Mapping for getting all participants
      * @return a list of all participants
+     * @param eventId the event that the participant belongs to
      */
     @GetMapping(path = { "", "/" })
-    public List<Participant> getAll() {
+    public List<Participant> getAll(@PathVariable("eventId") long eventId) {
         return repo.findAll();
     }
 
@@ -71,13 +82,16 @@ public class ParticipantController {
      * Mapping for changing a Participant
      * @param participant the updated participant to change to
      * @param id the id of the participant to change
+     * @param eventId the event that the participant belongs to
      * @return the participant changed or a bad request if no
      * participant can be found
      */
     @PutMapping("/{id}")
     public ResponseEntity<Participant> changeParticipant
-    (@RequestBody Participant participant, @PathVariable long id) {
-        if (id < 0 || !repo.existsById(id)) {
+    (@RequestBody Participant participant, @PathVariable long id,
+        @PathVariable("eventId") long eventId) {
+        if (id < 0 || !repo.existsById(id)
+                || participant.getEvent().getId() != eventId) {
             return ResponseEntity.badRequest().build();
         }
         return repo.findById(id)
@@ -87,18 +101,22 @@ public class ParticipantController {
     /**
      * Mapping for deleting a participant
      * @param id the id of the participant to remove
+     * @param eventId the event that the participant belongs to
      * @return the participant that was removed or bad
      * request if no participant was found
      */
     @DeleteMapping("/{id}")
     public ResponseEntity<Participant>
-        removeParticipant(@PathVariable long id) {
+        removeParticipant(@PathVariable long id,
+                          @PathVariable("eventId") long eventId) {
         if (id < 0 || !repo.existsById(id)) {
             return ResponseEntity.badRequest().build();
         }
-        ResponseEntity<Participant> participant =
-                ResponseEntity.ok(repo.findById(id).get());
+        Participant participant = repo.findById(id).get();
+
+        if (participant.getEvent().getId() != eventId)
+            return ResponseEntity.badRequest().build();
         repo.deleteById(id);
-        return participant;
+        return ResponseEntity.ok(participant);
     }
 }
